+++ src/Compiler/Dialects/Common/Type/DeleteCompiler.php
@@ -2,10 +2,11 @@
 
 namespace Somnambulist\Components\QueryBuilder\Compiler\Dialects\Common\Type;
 
+use Somnambulist\Components\QueryBuilder\Executors\ExecutableDeleteQuery;
 use Somnambulist\Components\QueryBuilder\Query\Type\DeleteQuery;
 
 class DeleteCompiler extends AbstractQueryCompiler
 {
     protected array $order = ['comment', 'with', 'delete', 'modifier', 'from', 'where', 'epilog'];
-    protected array $supports = [DeleteQuery::class];
+    protected array $supports = [DeleteQuery::class, ExecutableDeleteQuery::class];
 }