+++ src/Compiler/Dialects/Mysql/CompilerConfigurator.php
@@ -43,6 +43,7 @@ use Somnambulist\Components\QueryBuilder\Compiler\Dialects\Common\Expressions\Wh
 use Somnambulist\Components\QueryBuilder\Compiler\Dialects\Common\Expressions\WindowClauseCompiler;
 use Somnambulist\Components\QueryBuilder\Compiler\Dialects\Common\Expressions\WindowCompiler;
 use Somnambulist\Components\QueryBuilder\Compiler\Dialects\Common\Expressions\WithCompiler;
+use Somnambulist\Components\QueryBuilder\Compiler\Dialects\Common\Listeners\QuoteIdentifiers;
 use Somnambulist\Components\QueryBuilder\Compiler\Dialects\Common\Listeners\StripAliasesFromConditions;
 use Somnambulist\Components\QueryBuilder\Compiler\Dialects\Common\Listeners\StripAliasesFromDeleteFrom;
 use Somnambulist\Components\QueryBuilder\Compiler\Dialects\Common\Listeners\WrapUnionSelectClauses;
@@ -84,7 +85,7 @@ class CompilerConfigurator
     {
         return [
             PreDeleteQueryCompile::class => [
-                $qi = [new IdentifierQuoter('`', '`'), 'quote'],
+                $qi = new QuoteIdentifiers(new IdentifierQuoter('`', '`')),
                 new StripAliasesFromDeleteFrom(),
                 $sa = new StripAliasesFromConditions(),
             ],