+++ src/Executors/ExecutableSelectQuery.php
@@ -1,0 +1,11 @@
+<?php declare(strict_types=1);
+
+namespace Somnambulist\Components\QueryBuilder\Executors;
+
+use Somnambulist\Components\QueryBuilder\Executors\Behaviours\ExecutableQuery;
+use Somnambulist\Components\QueryBuilder\Query\Type\SelectQuery;
+
+class ExecutableSelectQuery extends SelectQuery
+{
+    use ExecutableQuery;
+}