+++ src/TypeCasterManager.php
@@ -4,6 +4,13 @@ namespace Somnambulist\Components\QueryBuilder;
 
 use Somnambulist\Components\QueryBuilder\Exceptions\TypeCasterException;
 
+/**
+ * Allows query values to be cast to something that can be understood by the DB server.
+ *
+ * Owing to how this is needed to be called during query building / compiling, it must be static.
+ * A type caster is unique for a given database driver implementation. i.e. one is needed for
+ * Doctrine, another for PDO, another for Laminas etc.
+ */
 class TypeCasterManager
 {
     private static ?TypeCaster $instance = null;
@@ -11,6 +18,11 @@ class TypeCasterManager
     public static function register(TypeCaster $caster): void
     {
         self::$instance = $caster;
+    }
+
+    public static function isRegistered(): bool
+    {
+        return self::$instance instanceof TypeCaster;
     }
 
     public static function instance(): TypeCaster